+++ CHANGELOG.md
@@ -7,7 +7,7 @@
    `setextHeading` for Setext headings
    [PR#63](https://github.com/chenzhiguang/dart_markdown/pull/63).
 3. **BREAKING** Use dedicated name `fencedBlockquote` for fenced blockquote
-   [PR#64](https://github.com/chenzhiguang/dart_markdown/pull/64).
+   [PR#65](https://github.com/chenzhiguang/dart_markdown/pull/65).
 
 ## 3.0.0