+++ CHANGELOG.md
@@ -10,7 +10,7 @@
    [PR#65](https://github.com/chenzhiguang/dart_markdown/pull/65).
 4. **BREAKING** Use dedicated name `autolinkExtension` for autolink extension
    and `autolink` for autolinks
-   [PR#67](https://github.com/chenzhiguang/dart_markdown/pull/67).
+   [PR#66](https://github.com/chenzhiguang/dart_markdown/pull/66).
 
 ## 3.0.0